pyproject.toml
[tool.poetry]
name = "swapy"
version = "0.1.0"
description = ""
authors = ["robgriffin247 <robgriffinwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
dagster = "^1.7.10"
dagster-webserver = "^1.7.10"
httpx = "^0.27.0"
dagster-duckdb = "^0.23.10"
dagster-duckdb-pandas = "^0.23.10"
dbt = "^1.0.0.37.22"
dbt-core = "^1.8.2"
dbt-duckdb = "^1.8.1"
dagster-dbt = "^0.23.10"
streamlit = "^1.37.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
module_name = "dags"